requirements.txt
glfw
moderngl
numpy
Pillow
